pyproject.toml
[project]
name = "project-n8tah"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.22.1",
    "google-generativeai>=0.8.6",
    "litellm>=1.81.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "supermemory>=3.18.0",
]
